Move CopyableThreadContextElement to common
