+++ kotlinx-coroutines-core/jsAndWasmShared/src/CoroutineContext.kt
@@ -6,16 +6,7 @@ import kotlin.coroutines.*
 internal actual val DefaultDelay: Delay
     get() = Dispatchers.Default as Delay
 
-public actual fun CoroutineScope.newCoroutineContext(context: CoroutineContext): CoroutineContext {
-    val combined = coroutineContext + context
-    return if (combined !== Dispatchers.Default && combined[ContinuationInterceptor] == null)
-        combined + Dispatchers.Default else combined
-}
-
-public actual fun CoroutineContext.newCoroutineContext(addedContext: CoroutineContext): CoroutineContext {
-    return this + addedContext
-}
-
 // No debugging facilities on Wasm and JS
 internal actual fun Continuation<*>.toDebugString(): String = toString()
 internal actual val CoroutineContext.coroutineName: String? get() = null // not supported on Wasm and JS
+internal actual fun wrapContextWithDebug(context: CoroutineContext): CoroutineContext = context
