+++ kotlinx-coroutines-core/native/src/CoroutineContext.kt
@@ -28,16 +28,7 @@ internal expect fun createDefaultDispatcher(): CoroutineDispatcher
 @PublishedApi
 internal actual val DefaultDelay: Delay = DefaultExecutor
 
-public actual fun CoroutineScope.newCoroutineContext(context: CoroutineContext): CoroutineContext {
-    val combined = coroutineContext + context
-    return if (combined !== Dispatchers.Default && combined[ContinuationInterceptor] == null)
-        combined + Dispatchers.Default else combined
-}
-
-public actual fun CoroutineContext.newCoroutineContext(addedContext: CoroutineContext): CoroutineContext {
-    return this + addedContext
-}
-
 // No debugging facilities on native
 internal actual fun Continuation<*>.toDebugString(): String = toString()
 internal actual val CoroutineContext.coroutineName: String? get() = null // not supported on native
+internal actual fun wrapContextWithDebug(context: CoroutineContext): CoroutineContext = context
